# Acquire sample rows for root partition to build extend statistics