--- a/src/backend/commands/analyzeutils.c
+++ b/src/backend/commands/analyzeutils.c
@@ -1029,9 +1029,10 @@ getBucketSizes(const HeapTuple *heaptupleStats, const float4 *relTuples, int nPa
  *	needs_sample() -- checks if the analyze requires sampling the actual data
  */
 bool
-needs_sample(VacAttrStats **vacattrstats, int attr_cnt)
+needs_sample(Relation rel, VacAttrStats **vacattrstats, int attr_cnt)
 {
 	Assert(vacattrstats != NULL);
+	List *statext_oids;
 	int			i;
 
 	for (i = 0; i < attr_cnt; i++)
@@ -1040,6 +1041,15 @@ needs_sample(VacAttrStats **vacattrstats, int attr_cnt)
 		if (!vacattrstats[i]->merge_stats)
 			return true;
 	}
+
+	/* we must acquire sample rows to build extend statisics */
+	statext_oids = RelationGetStatExtList(rel);
+	if (statext_oids != NIL)
+	{
+		list_free(statext_oids);
+		return true;
+	}
+
 	return false;
 }