This pr contains the following modifications：
- fixed issue https://github.com/greenplum-db/gpdb/issues/13872.
We used to not acquire sample rows for root partition when the following conditions were held for all attributes.
```c
	if (get_rel_relkind(attr->attrelid) == RELKIND_PARTITIONED_TABLE &&
		!get_rel_relispartition(attr->attrelid) &&
		leaf_parts_analyzed(stats->attr->attrelid, InvalidOid, va_cols, stats->elevel) &&
		((!OidIsValid(eqopr)) || op_hashjoinable(eqopr, stats->attrtypid)))
```
It works fine when we just have per-column-statistics, we can merge the statistics of leaf partitions into statistics of the root partition. But it can't work with extended statistics. We must acquire sample rows for root partitions to build extended statistics. 

- cherry-picked 8356f5d22db2558895226deeea021b950cb7d839 from upstream.

- cherry-picked ba685d67090dd7ea905c7c052ab9798182254e96 from upstream.

- cherry-picked 8eb8a3b68b75147b5ebc6fca95fe1c4f4c3d3460 from upstream.

## Here are some reminders before you submit the pull request
- [ ] Add tests for the change
- [ ] Document changes
- [ ] Communicate in the mailing list if needed
- [ ] Pass `make installcheck`
- [ ] Review a PR in return to support the community