@ashwinstar 2022-11-15
Please can you describe the problem a little more, just reading the PR description (not reading code) not clear what does it mean by acquire sample rows for root partition? (like root doesn't contain data anyways) Does it mean for full hierarchy?