@QingMa-gp 2022-11-15
> Please can you describe the problem a little more, just reading the PR description (not reading code) not clear what does it mean by acquire sample rows for root partition? (like root doesn't contain data anyways) Does it mean for full hierarchy?

@ashwinstar You are right, it means we need to acquire sample rows from inheritance tree to build extend statistics for root partition, but the current code skips `acquire_inherited_sample_rows()` for root partition when [certain conditions](https://github.com/greenplum-db/gpdb/blob/main/src/backend/commands/analyze.c#L683-L716) are met.