@QingMa-gp 2022-11-16
Hi, @chrishajas.

> If I understand this correctly, it gets rid of the entirety of the merging leaf stats/hyperloglog logic? I don't think we can do this like this PR. With many partitions, analyze can take a VERY long time, and updating 1 partition would mean that all tables in the partition need to be re-sampled.
>
> If possible we want to maintain the merging leaf stats architecture here. I'm not exactly sure what that means for extended statistics, likely it means adapting the HLL logic to support merging extended statistics.

This pr doesn't change the merging leaf stats/HLL logic, it just gets rid of the logic that skips acquiring sample rows for root partition. 
  
> Does planner use the partition statistics directly or does it only use leaf partitions?
Planner has a guc `gp_statistics_pullup_from_child_partition` to control this, it's disabled by default.  If it is enabled, we try to use the largest child's statistics. Otherwise, we use the root partition statistics directly.