@ashwinstar 2022-11-16
Seems this PR would be performance sensitive, so do we have performance numbers to understand the impact on analyze from the change (lets say with many small partitions or with some large partitions and such)