@chrishajas 2022-11-17
Would we be able to modify the logic to only do sampling if extended stats are present for root partitions? Possibly using something like `stats = fetch_statentries_for_relation(pg_stext, RelationGetRelid(onerel));`

I synced with @kainwen yesterday, and there are efforts to improve analyze performance (https://github.com/greenplum-db/gpdb/pull/14416) after which we can re-evaluate merging vs sampling and how we want to handle either merging or sampling for extended stats.