@QingMa-gp 2022-11-18
>Seems this PR would be performance sensitive, so do we have performance numbers to understand the impact on analyze from the change (lets say with many small partitions or with some large partitions and such)

I did some quick tests, the results show that this patch has less impact on the performance of analyze with many small partitions, but it almost doubles the time of analysis with some large partitions.

The test steps and results are pasted below:
```SQL
-- table name : test_<storage>_npart
-- each table has one hundred million rows

create table test_heap_100(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int)
distributed by (a) partition by range (b) ( start (0) end (1000) every (1));

insert into test_heap_100 select i, i % 1000, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;
create table test_ao_100(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int) with (appendonly=true, orientation=row)
distributed by (a) partition by range (b) ( start (0) end (1000) every (1));

insert into test_ao_100 select i, i % 1000, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;

create table test_aoco_100(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int) with (appendonly=true, orientation=column)
distributed by (a) partition by range (b) ( start (0) end (1000) every (1));

insert into test_aoco_100 select i, i % 1000, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;

create table test_heap_10(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int)
distributed by (a) partition by range (b) ( start (0) end (10) every (1));

insert into test_heap_10 select i, i % 10, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;

create table test_ao_10(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int) with (appendonly=true, orientation=row)
distributed by (a) partition by range (b) ( start (0) end (10) every (1));

insert into test_ao_10 select i, i % 10, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;
create table test_aoco_10(a int, b int, c int, d int, e int, f int, g int, h int, i int, j int, k int, l int, m int, n int, o int, p int, q int, r int, s int, t int, u int, v int, w int, x int, y int, z int) with (appendonly=true, orientation=column)
distributed by (a) partition by range (b) ( start (0) end (10) every (1));

insert into test_aoco_10 select i, i % 10, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i, i from generate_series(1,100000000)i;

-- without patch

postgres=# analyze test_heap_100;
ANALYZE
Time: 1042752.262 ms (17:22.752)
postgres=# analyze test_ao_100;
ANALYZE
Time: 1061616.591 ms (17:41.617)
postgres=# analyze test_aoco;
ANALYZE
Time: 1089938.174 ms (18:09.938)
postgres=# analyze test_heap_10;
ANALYZE
Time: 34690.224 ms (00:34.690)
postgres=# analyze test_ao_10;
ANALYZE
Time: 49969.162 ms (00:49.969)
postgres=# analyze test_aoco_10;
ANALYZE
Time: 68505.805 ms (01:08.506)

-- with patch
postgres=# analyze test_heap_100;
ANALYZE
Time: 1070710.749 ms (17:50.711)
postgres=# analyze test_ao_100;
ANALYZE
Time: 1125751.630 ms (18:45.752)
postgres=# analyze test_aoco_100;
ANALYZE
Time: 1170464.588 ms (19:30.465)
postgres=# analyze test_heap_10;
ANALYZE
Time: 37834.044 ms (00:37.834)
postgres=# analyze test_ao_10;
ANALYZE
Time: 85895.867 ms (01:25.896)
postgres=# analyze test_aoco_10;
ANALYZE
Time: 119500.059 ms (01:59.500)
```